add threads examples

+++ Endgame.h
@@ -11,27 +11,27 @@ bool getWinner(entities &allEntities)
         for (auto worker : allEntities.workers) race_frecuency[worker->RACE()]++;
         for (auto soldier : allEntities.soldiers) race_frecuency[soldier->RACE()]++;
         for (auto frecuency : race_frecuency) if (frecuency.second == 0) {
-            for (int i=0;i<allEntities.bases.size();i++) if (frecuency.first==allEntities.bases[i]->RACE()) {
+            for (int i=0;i<(int)allEntities.bases.size();i++) if (frecuency.first==allEntities.bases[i]->RACE()) {
                 allEntities.bases[i]->erase();
                 allEntities.bases.erase(allEntities.bases.begin()+i);
                 i=0;
             }
-            for (int i=0;i<allEntities.worker_generators.size();i++) if (frecuency.first==allEntities.worker_generators[i]->RACE()) {
+            for (int i=0;i<(int)allEntities.worker_generators.size();i++) if (frecuency.first==allEntities.worker_generators[i]->RACE()) {
                 allEntities.worker_generators[i]->erase();
                 allEntities.worker_generators.erase(allEntities.worker_generators.begin()+i);
                 i=0;
             }
-            for (int i=0;i<allEntities.soldier_generators.size();i++) if (frecuency.first==allEntities.soldier_generators[i]->RACE()) {
+            for (int i=0;i<(int)allEntities.soldier_generators.size();i++) if (frecuency.first==allEntities.soldier_generators[i]->RACE()) {
                 allEntities.soldier_generators[i]->erase();
                 allEntities.soldier_generators.erase(allEntities.soldier_generators.begin()+i);
                 i=0;
             }
-            for (int i=0;i<allEntities.skills_structures.size();i++) if (frecuency.first==allEntities.skills_structures[i]->RACE()) {
+            for (int i=0;i<(int)allEntities.skills_structures.size();i++) if (frecuency.first==allEntities.skills_structures[i]->RACE()) {
                 allEntities.skills_structures[i]->erase();
                 allEntities.skills_structures.erase(allEntities.skills_structures.begin()+i);
                 i=0;
             }
-            for (int i=0;i<allEntities.kamikazes.size();i++) if (frecuency.first==allEntities.kamikazes[i]->RACE()) {
+            for (int i=0;i<(int)allEntities.kamikazes.size();i++) if (frecuency.first==allEntities.kamikazes[i]->RACE()) {
                 allEntities.kamikazes[i]->erase();
                 allEntities.kamikazes.erase(allEntities.kamikazes.begin()+i);
                 i=0;
